moving UT to BatchSchemaTests.cs
